Basic implementation for StartExecution manager/worker
